pyproject.toml
[project]
name = "competitor-intel"
version = "1.0.0"
description = "AI-Powered Competitor Intelligence Engine"
requires-python = ">=3.10,<3.14"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn>=0.23.0,<0.35",
    "pydantic>=2.10.4,<2.13",
    "playwright>=1.49.1",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.3.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-groq>=0.2.0",
    "requests>=2.32.3",
    "pinecone>=5.0.0",
    "sqlalchemy>=2.0.36",
    "tiktoken>=0.8.0",
    "reportlab>=4.2.5",
    "Pillow>=11.1.0",
    "streamlit>=1.35.0,<2",
    "requests>=2.32.3",
    "aiofiles>=24.1.0",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.0.0",
    "duckduckgo-search>=7.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
no-build-isolation = true